VORTEX-6077: Fixing layout of mini style panel.
